+++ ingot-service/ingot-pms/ingot-pms-provider/src/main/java/com/ingot/cloud/pms/service/domain/SysAuthorityService.java
@@ -42,5 +42,5 @@ public interface SysAuthorityService extends BaseService<SysAuthority> {
      *
      * @param id 权限ID
      */
-    void removeAuthorityById(long id);
+    void removeAuthorityById(int id);
 }
